Code refactoring and logic fixes, as well as an improved API request with a timeout based on the remaining time between the current track and the upcoming synced lyric position. This feature will prevent many unnecessary API requests to the Spotify server and locally.
Code refactoring and logic fixes, as well as an improved API request with a timeout based on the remaining time between the current track and the upcoming synced lyric position. This feature will prevent many unnecessary API requests to the Spotify server and locally.

Code refactoring and logic fixes, as well as an improved API request with a timeout based on the remaining time between the current track and the upcoming synced lyric position. This feature will prevent many unnecessary API requests to the Spotify server and locally.